Move `adjustNoSuchWallet` and `ErrNoSuchWallet`
… to the `Cardano.Wallet.DB.WalletState` module.
