+++ lib/core/test/unit/Cardano/Wallet/DB/StateMachine.hs
@@ -61,7 +61,6 @@ import Cardano.Address.Script
 import Cardano.Wallet.DB
     ( DBLayer (..)
     , ErrNoSuchTransaction (..)
-    , ErrNoSuchWallet (..)
     , ErrPutLocalTxSubmission (..)
     , ErrRemoveTx (..)
     , ErrWalletAlreadyExists (..)
@@ -99,6 +98,8 @@ import Cardano.Wallet.DB.Pure.Implementation
     , mRollbackTo
     , mUpdatePendingTxForExpiry
     )
+import Cardano.Wallet.DB.WalletState
+    ( ErrNoSuchWallet (..) )
 import Cardano.Wallet.DummyTarget.Primitive.Types
     ( dummyGenesisParameters, dummyTimeInterpreter )
 import Cardano.Wallet.Primitive.AddressDerivation
